some further work

+++ Libraries/01-Shared/Elysium.Core.Globalization/CalendarAlgorithmType.hpp
@@ -13,17 +13,17 @@ Copyright (c) waYne (CAM). All rights reserved.
 #endif
 
 #ifndef ELYSIUM_CORE_PRIMITIVES
-#include "../Elysium.Core/Primitives.hpp"
+#include "../Elysium.Core.Template/Primitives.hpp"
 #endif
 
 #ifndef ELYSIUM_CORE_SYSTEM
-#include "../Elysium.Core/System.hpp"
+#include "../Elysium.Core.Template/System.hpp"
 #endif
 
 namespace Elysium::Core::Globalization
 {
 #if defined(ELYSIUM_CORE_OS_WINDOWS)
-	enum class CalendarAlgorithmType : Elysium::Core::uint32_t
+	enum class CalendarAlgorithmType : Elysium::Core::uint8_t
 #elif defined(ELYSIUM_CORE_OS_ANDROID)
 	enum class CalendarAlgorithmType
 #else
